Generalized Conditional Functional Principal Components Analysis (GC-FPCA)

This is the main function for the GCFPCA package.

Usage

gc_fpca(
  formula,
  data,
  binwidth = 10,
  family = "gaussian",
  pve = NULL,
  npc = NULL,
  periodicity = FALSE,
  ...
)

Arguments

formula

A formula specifying the GLMM model.

data

A data frame with columns: id (subject), index (time or index), and Y (outcome variable).

binwidth

Numeric; controls the width of the bins for local fitting. Defaults to 10.

family

A family object for specifying the distribution and link function for the GLMM. Defaults to binomial.

pve

Numeric; proportion of variance explained, used to choose the number of principal components unless npc is specified.

npc

Numeric; number of smooth PCs to extract. If NULL, npc is chosen based on pve.

...

Additional arguments passed to the mgcv::bam function.

Value

An object of class gc_fpca containing:

betaHat

Estimated fixed effects coefficient functions

betaHat.sd

Estimated fixed effects coefficient standard deviation along the functional domain

efunctions

\(K \times npc\) matrix of estimated FPC basis functions.

scores

\(I \times npc\) matrix of estimated FPC scores.

mu

Estimated population-level mean function.

Yhat

FPC approximation of subject-specific means.

family

The family used for GLMM.

Examples

# Simulate data and fit GC-FPCA
df_list <- gen_data(N = 100, J = 50, run_num = 1)
#> Error in gen_data(N = 100, J = 50, run_num = 1): could not find function "gen_data"
gcfpca_mod <- gc_fpca(df_list, bin_width = 10, family = binomial())
#> Error: object 'df_list' not found