Simulate longitudinal functional response Y (matrix) and a scalar predictor X (vector)

Simulate longitudinal functional response Y (matrix) and a scalar predictor X (vector)

Usage

gcfpca_simu(
  family = "gaussian",
  I = 100,
  K = 100,
  beta0_true,
  beta1_true,
  fe_case = 1,
  re_case = 1
)

Arguments

family

distribution of longitudinal functional data, including "gaussian", "binomial", "poisson".

I

number of subjects.

K

number of grid points on the functional domain.

beta0_true

true intercept function.

beta1_true

true fixed effect function.

fe_case

Takes on values of 1 or 2. If case = 1, true fixed effect coefficient is generated from a Bernoulli(0.5) distribution . If case = 2 true fixed effect, coefficient is from a N(0,1) distribution.

re_case

Takes on values of 1 or 2. If case = 1 true eigenfunctions are based on alternating sine and cosines. If case = 2 true eigenfunctions are based on sqrt functions.

Value

a data frame containing generated predictors and longitudinal functional outcomes.